requirements.txt
pandas
request
dash
arrow
plotly